support/requirements.txt
# Dependencias para cargar invitados a Supabase
pandas>=2.0.0
openpyxl>=3.1.0
supabase>=2.0.0
python-dotenv>=1.0.0
